+++ packages/xd-crossword-tools-parser/src/parser/xdparser.migration.test.ts
@@ -8,11 +8,8 @@ it("converts a v1 implicit xd to explicit under the hood", () => {
 })
 
 it("correctly knows whether to do the transition", () => {
-  const xd = readFileSync("./packages/xd-crossword-tools-parser/src/xdparser2/inputs/alpha-bits.xd", "utf8")
-  const explicitXD = readFileSync(
-    "./packages/xd-crossword-tools-parser/src/xdparser2/packages/xd-crossword-tools/tests/puz/explicit-alpha-bits.xd",
-    "utf8"
-  )
+  const xd = readFileSync("./packages/xd-crossword-tools-parser/src/parser/inputs/alpha-bits.xd", "utf8")
+  const explicitXD = readFileSync("./packages/xd-crossword-tools-parser/src/parser/outputs/explicit-alpha-bits.xd", "utf8")
 
   expect(shouldConvertToExplicitHeaders(xd)).toBeTruthy()
   expect(shouldConvertToExplicitHeaders(explicitXD)).toBeFalsy()